Fix signed-unsigned-warning in test

+++ source/tests/fiblib-test/fibonacci_test.cpp
@@ -13,9 +13,9 @@ TEST_F(fibonacci_test, CheckSomeResults)
 {
     fiblib::Fibonacci fib;
 
-    EXPECT_EQ( 0, fib( 0));
-    EXPECT_EQ( 1, fib( 1));
-    EXPECT_EQ( 1, fib( 2));
-    EXPECT_EQ(21, fib( 8));
+    EXPECT_EQ((unsigned int) 0, fib(0));
+    EXPECT_EQ((unsigned int) 1, fib(1));
+    EXPECT_EQ((unsigned int) 1, fib(2));
+    EXPECT_EQ((unsigned int)21, fib(8));
     // ...
 }
